pyproject.toml
[tool.poetry]
name = "ekubo"
version = "0.1.0"
description = ""
authors = ["Clément Walter <clement0walter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.1,<3.12"
pandas = "^2.1.1"
requests = "^2.31.0"
starknet-py = "^0.18.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
